+++ PiCamera/pi_camera.hpp
@@ -232,7 +232,7 @@ extern "C"
 
 	PI_CAMERA_API_EXPORT AL::uint8 PI_CAMERA_API_CALL pi_camera_open(pi_camera** camera);
 	PI_CAMERA_API_EXPORT AL::uint8 PI_CAMERA_API_CALL pi_camera_open_remote(pi_camera** camera, const char* remote_host, AL::uint16 remote_port);
-	PI_CAMERA_API_EXPORT AL::uint8 PI_CAMERA_API_CALL pi_camera_open_service(pi_camera** camera, const char* local_host, AL::uint16 local_port, AL::size_t max_connections);
+	PI_CAMERA_API_EXPORT AL::uint8 PI_CAMERA_API_CALL pi_camera_open_service(pi_camera** camera, const char* local_host, AL::uint16 local_port, AL::uint32 max_connections);
 	PI_CAMERA_API_EXPORT void      PI_CAMERA_API_CALL pi_camera_close(pi_camera* camera);
 
 	PI_CAMERA_API_EXPORT AL::uint8 PI_CAMERA_API_CALL pi_camera_is_busy(pi_camera* camera, bool* value);
